Docker/requirements.txt
scanpy>=1.9.6
scvi-tools>=1.1.3
anndata>=0.9
scikit-misc
numpy
pandas
scipy
tqdm
h5py
